requirements.txt
chainlit
python-dotenv
google-generativeai
-e .
